+++ src/main/java/com/api/payment/payment_api_service/controller/dto/PostPaymentResponse.java
@@ -5,12 +5,12 @@ import com.api.payment.payment_api_service.domain.PaymentMethod;
 import com.api.payment.payment_api_service.domain.PaymentStatusInit;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 
 public record PostPaymentResponse(
         PaymentMethod paymentMethod,
-        BigDecimal paymentValue,
+        BigInteger paymentValue,
         Integer debtCode,
         PayerType payerType,
         String payerId,
